requirements.txt
# LLM APIs (choose one or both)
anthropic>=0.7.7
openai>=1.3.7
google-genai

# Web scraping & HTTP
requests>=2.31.0
beautifulsoup4>=4.12.2
lxml>=4.9.3

# Utilities
python-dotenv>=1.0.0